DAY3-8/requirements.txt
websockets==15.0.1
python-dotenv==1.2.1
cryptography==46.0.5